cut borders and formatting
