+++ src/ui/menus/TopBar/SubMenus/useBorders.ts
@@ -1,8 +1,8 @@
+import { clearBordersAction } from '../../../../core/actions/clearBordersAction';
 import { Border, BorderType } from '../../../../core/gridDB/gridTypes';
 import { localFiles } from '../../../../core/gridDB/localFiles';
 import { Sheet } from '../../../../core/gridDB/Sheet';
 import { PixiApp } from '../../../../core/gridGL/pixiApp/PixiApp';
-import { Coordinate } from '../../../../core/gridGL/types/size';
 import { useGetSelection } from './useGetSelection';
 
 export interface ChangeBorder {
@@ -131,67 +131,7 @@ export const useBorders = (sheet: Sheet, app: PixiApp): IResults => {
   };
 
   const clearBorders = (args?: { create_transaction?: boolean }): void => {
-    const borderUpdate: Border[] = [];
-    const borderDelete: Coordinate[] = [];
-    for (let y = start.y; y <= end.y; y++) {
-      for (let x = start.x; x <= end.x; x++) {
-        const border = sheet.borders.get(x, y);
-        if (border) {
-          borderDelete.push({ x, y });
-        }
-        if (x === end.x) {
-          const border = sheet.borders.get(x + 1, y);
-          if (border?.vertical) {
-            if (!border.horizontal) {
-              borderDelete.push({ x: x + 1, y });
-            } else {
-              borderUpdate.push({ ...border, vertical: undefined });
-            }
-          }
-        }
-        if (y === end.y) {
-          const border = sheet.borders.get(x, y + 1);
-          if (border?.horizontal) {
-            if (!border.vertical) {
-              borderDelete.push({ x, y: y + 1 });
-            } else {
-              borderUpdate.push({ ...border, horizontal: undefined });
-            }
-          }
-        }
-      }
-    }
-
-    // create transaction to update borders
-    args?.create_transaction ?? sheet_controller.start_transaction();
-    if (borderDelete.length) {
-      borderDelete.forEach((border_coord) => {
-        sheet_controller.execute_statement({
-          type: 'SET_BORDER',
-          data: {
-            position: [border_coord.x, border_coord.y],
-            border: undefined,
-          },
-        });
-      });
-    }
-    if (borderUpdate.length) {
-      borderUpdate.forEach((border) => {
-        sheet_controller.execute_statement({
-          type: 'SET_BORDER',
-          data: {
-            position: [border.x, border.y],
-            border: border,
-          },
-        });
-      });
-    }
-    args?.create_transaction ?? sheet_controller.end_transaction();
-
-    app.cells.dirty = true;
-    app.quadrants.quadrantChanged({ range: { start, end } });
-
-    localFiles.saveLastLocal(sheet.export_file());
+    clearBordersAction({ sheet_controller, start, end, create_transaction: args?.create_transaction });
   };
 
   return {
